Add files via upload
This report uses the approach of Price prediction-based trading. First a model is built to predict the prices
for the coming day. Using the prediction a trading rule is built which chooses to buy/sell/hold based on the
current and predicted values. We then back-test the built trading rule on historical data.